+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Calendar Versioning](https://calver.org/) of
 the following form: YYYY.0M.0D.
 
+## 2025.11.08
+
+### Changed
+
+- Split isValidTDFile() into isValidHeader() and isValidDataRow()
+
+- Modified readFromFile() to better check for errors and integrate contract checking methods (isTxt(), isValidHeader(), isValidDataRow())
+
 ## 2025.11.07
 
 ### Added